backend/requirements.txt
Flask==2.0.2
requests==2.28.1 
flask_cors==3.0.10
azure-cognitiveservices-speech